Extract tests for isRegExp to own file
